+++ DesignerTest/AddDesignTests.cs
@@ -7,9 +7,9 @@ using ServicesTest;
 
 namespace DesignerTest {
     public class AddDesignDatabaseTestsStaticMethods : DatabaseTest {
-        private static readonly Room Room1 = new Room.FromDimensions("TestRoom1", 1, 1);
-        private static readonly Room Room2 = new Room.FromDimensions("TestRoom2", 2, 4);
-        private static readonly Room Room3 = new Room.FromDimensions("TestRoom3", 2, 5);
+        private static readonly Room Room1 = Room.FromDimensions("TestRoom1", 1, 1);
+        private static readonly Room Room2 = Room.FromDimensions("TestRoom2", 2, 4);
+        private static readonly Room Room3 = Room.FromDimensions("TestRoom3", 2, 5);
         
         protected override List<Room> Rooms => new List<Room> {Room1, Room2, Room3};
 
@@ -94,10 +94,10 @@ namespace DesignerTest {
         }
     }
 
-    public class AddDesignTestsInstace {
-        private static readonly Room Room1 = new Room.FromDimensions("TestRoom1", 1, 1);
-        private static readonly Room Room2 = new Room.FromDimensions("TestRoom2", 2, 4);
-        private static readonly Room Room3 = new Room.FromDimensions("TestRoom3", 2, 5);
+    public class AddDesignTestsInstace : DatabaseTest {
+        private static readonly Room Room1 = Room.FromDimensions("TestRoom1", 1, 1);
+        private static readonly Room Room2 = Room.FromDimensions("TestRoom2", 2, 4);
+        private static readonly Room Room3 = Room.FromDimensions("TestRoom3", 2, 5);
 
         private static readonly string TestName = "TestDesign";
