+++ DesignerTest/DesignerTest.csproj
@@ -17,4 +17,8 @@
       <ProjectReference Include="..\ServicesTest\ServicesTest.csproj" />
     </ItemGroup>
 
+    <ItemGroup>
+      <Folder Include="ViewModel" />
+    </ItemGroup>
+
 </Project>
